Create delete salary

+++ src/main/java/com/manage/salary/salary/Salary.java
@@ -26,7 +26,7 @@ public class Salary {
     private Double netMonth;
 
     @Column(name = "tax", nullable = false)
-    private Integer tax;
+    private Double tax;
 
     @ManyToOne(fetch = FetchType.LAZY,
             cascade = {CascadeType.DETACH, CascadeType.MERGE,
